refactor(conversation):WB-3957 add useToolbarVisibility hook
